+++ code/frontend/src/components/sign.js
@@ -1,0 +1,83 @@
+
+import axios from 'axios'
+
+export default {
+  getCookieName(k) {
+    return (document.cookie.match("(^|; )" + k + "=([^;]*)") || 0)[2];
+  },
+  in(signInData, isSaler) {
+    let url = ""
+    if (isSaler) {
+      url += "/api/saler/login";
+    } else {
+      url += "/api/customer/login";
+    }
+    axios
+      .post(url, signInData, {
+        headers: {
+          "X-CSRFToken": this.getCookieName("csrftoken"),
+        },
+      })
+      .then((response) => {
+        console.log(response);
+        confirm("Sign in successfully!");
+        let data = response.data;
+
+        localStorage.setItem("name", data.name);
+        localStorage.setItem("csrftoken", this.getCookieName("csrftoken"))
+        if (this.isSaler) {
+          localStorage.setItem("isSaler", "y");
+          window.location.href = "/salerinfo";
+        } else {
+          localStorage.setItem("isSaler", "n");
+          window.location.href = "/customerinfo";
+        }
+      })
+      .catch((error) => {
+        alert(error);
+      });
+  },
+  out() {
+    try {
+      let id = localStorage.getItem("isSaler");
+      if (id === "n") {
+        axios.get("/api/customer/logout");
+      } else if (id === "y") {
+        axios.get("/api/saler/logout");
+      }
+    } catch (error) {
+      console.log(error);
+      alert(error);
+      return;
+    }
+    localStorage.clear();
+    window.location.href = "/";
+  },
+  up(signUpData, isSaler, file) {
+    const formData = new FormData();
+
+    Object.keys(signUpData).forEach((key) =>
+      formData.append(key, this.signUpData[key])
+    );
+    let url = "";
+    if (isSaler) {
+      url += "/api/saler/signup";
+    } else {
+      url += "/api/customer/signup";
+      formData.append("self_pics", file);
+    }
+    axios
+      .post(url, formData, {
+        headers: {
+          "Content-Type": "multipart/form-data",
+        },
+      })
+      .then((response) => {
+        console.log(response);
+        confirm("Sign up successfully!");
+      })
+      .catch((error) => {
+        alert(error);
+      });
+  }
+}
