+++ code/frontend/src/pages/Home/main.js
@@ -1,10 +1,8 @@
 import Vue from 'vue'
 import App from './App.vue'
 import vuetify from '@/plugins/vuetify';
-import axios from 'axios'
 
 Vue.config.productionTip = false
-Vue.prototype.$http = axios
 
 new Vue({
   vuetify,
